Buttons should now looks like on a mockup.
